backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
orjson>=3.2.1
httpx>=0.23.0
starlette>=0.37.2
python-multipart>=0.0.7
fastapi-cli>=0.0.2
ujson>=4.0.1
typing-extensions>=4.8.0
email_validator>=2.0.0
pydantic>=1.7.4
jinja2>=2.11.2
h11>=0.8
anyio
typer>=0.12.3
dnspython>=2.0.0
annotated-types>=0.4.0
pydantic-core>=2.18.4
MarkupSafe>=2.0
exceptiongroup>=1.0.2
shellingham>=1.3.0
rich>=10.11.0
markdown-it-py>=2.2.0
pygments>=2.13.0
mdurl>=0.1
httptools>=0.1.0
python-dotenv
async-timeout>=4.0.3
asyncpg>=0.29.0
databases>=0.9.0
greenlet>=3.0.3
sqlalchemy>=2.0.30
bcrypt>=4.1.3
python-jose>=3.3.0
EOL
